pyproject.toml
[project]
name = "voice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "audioop-lts>=0.2.2",
    "fastapi[standard]>=0.117.1",
    "numpy>=2.3.3",
    "sounddevice>=0.5.2",
    "websocket-client>=1.7.0",
    "uvicorn[standard]>=0.37.0",
    "python-dotenv>=1.0.1",
]
